src/server/requirements.txt
beautifulsoup4==4.12.3
numpy==1.24.3
pandas==1.5.3
pymongo==4.7.2
python-dotenv==1.0.1
requests[security]
yfinance==0.2.38
openpyxl
fastapi
uvicorn
